moved pod files to resource/pod folder
