+++ src/core/cluster_checks.ts
@@ -7,7 +7,7 @@ import {patchInject} from './container_helper.js';
 import {SoloLogger} from './logging.js';
 import {inject, injectable} from 'tsyringe-neo';
 import {type K8} from './kube/k8.js';
-import {type Pod} from './kube/pod.js';
+import {type Pod} from './kube/resources/pod/pod.js';
 import {type IngressClass} from './kube/ingress_class.js';
 import {type V1Pod, type V1ConfigMap} from '@kubernetes/client-node';
