+++ src/core/kube/container_ref.ts
@@ -1,7 +1,7 @@
 /**
  * SPDX-License-Identifier: Apache-2.0
  */
-import {type PodRef} from './pod_ref.js';
+import {type PodRef} from './resources/pod/pod_ref.js';
 import {type ContainerName} from './container_name.js';
 import {NestedResourceRef} from './nested_resource_ref.js';
