+++ src/core/kube/k8_client.ts
@@ -19,14 +19,14 @@ import {K8ClientClusters} from './k8_client/k8_client_clusters.js';
 import {type Clusters} from './clusters.js';
 import {type ConfigMaps} from './config_maps.js';
 import {K8ClientConfigMaps} from './k8_client/k8_client_config_maps.js';
-import {type PodRef} from './pod_ref.js';
+import {type PodRef} from './resources/pod/pod_ref.js';
 import {type ContainerRef} from './container_ref.js';
 import {K8ClientContainers} from './k8_client/k8_client_containers.js';
 import {type Containers} from './containers.js';
 import {type Contexts} from './contexts.js';
 import {K8ClientContexts} from './k8_client/k8_client_contexts.js';
-import {K8ClientPods} from './k8_client/k8_client_pods.js';
-import {type Pods} from './pods.js';
+import {K8ClientPods} from './k8_client/resources/pod/k8_client_pods.js';
+import {type Pods} from './resources/pod/pods.js';
 import {K8ClientBase} from './k8_client/k8_client_base.js';
 import {type Services} from './services.js';
 import {K8ClientServices} from './k8_client/k8_client_services.js';
