+++ src/core/kube/k8_client/resources/pod/k8_client_pods.ts
@@ -11,22 +11,22 @@ import {
   V1ExecAction,
   V1PodSpec,
 } from '@kubernetes/client-node';
-import {type Pods} from '../pods.js';
-import {NamespaceName} from '../resources/namespace/namespace_name.js';
-import {PodRef} from '../pod_ref.js';
-import {type Pod} from '../pod.js';
+import {type Pods} from '../../../resources/pod/pods.js';
+import {NamespaceName} from '../../../resources/namespace/namespace_name.js';
+import {PodRef} from '../../../resources/pod/pod_ref.js';
+import {type Pod} from '../../../resources/pod/pod.js';
 import {K8ClientPod} from './k8_client_pod.js';
-import {Duration} from '../../time/duration.js';
-import {K8ClientBase} from './k8_client_base.js';
-import {MissingArgumentError, SoloError} from '../../errors.js';
-import * as constants from '../../constants.js';
-import {SoloLogger} from '../../logging.js';
+import {Duration} from '../../../../time/duration.js';
+import {K8ClientBase} from '../../k8_client_base.js';
+import {MissingArgumentError, SoloError} from '../../../../errors.js';
+import * as constants from '../../../../constants.js';
+import {SoloLogger} from '../../../../logging.js';
 import {container} from 'tsyringe-neo';
-import {type ContainerName} from '../container_name.js';
-import {PodName} from '../pod_name.js';
-import {KubeApiResponse} from '../kube_api_response.js';
-import {ResourceOperation} from '../resource_operation.js';
-import {ResourceType} from '../resource_type.js';
+import {type ContainerName} from '../../../container_name.js';
+import {PodName} from '../../../resources/pod/pod_name.js';
+import {KubeApiResponse} from '../../../kube_api_response.js';
+import {ResourceOperation} from '../../../resource_operation.js';
+import {ResourceType} from '../../../resource_type.js';
 
 export class K8ClientPods extends K8ClientBase implements Pods {
   private readonly logger: SoloLogger;
