+++ src/core/kube/resources/pod/pod_ref.ts
@@ -2,8 +2,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import {type PodName} from './pod_name.js';
-import {type NamespaceName} from './resources/namespace/namespace_name.js';
-import {ResourceRef} from './resource_ref.js';
+import {type NamespaceName} from '../namespace/namespace_name.js';
+import {ResourceRef} from '../../resource_ref.js';
 
 /**
  * Represents a Kubernetes pod reference which includes the namespace name and pod name.
