+++ src/core/kube/resources/pod/pods.ts
@@ -2,10 +2,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import {type V1Pod} from '@kubernetes/client-node';
-import {type NamespaceName} from './resources/namespace/namespace_name.js';
-import {type PodRef} from './pod_ref.js';
-import {type Pod} from './pod.js';
-import {type ContainerName} from './container_name.js';
+import {type NamespaceName} from '../namespace/namespace_name.js';
+import {type PodRef} from '../pod/pod_ref.js';
+import {type Pod} from '../pod/pod.js';
+import {type ContainerName} from '../../container_name.js';
 
 export interface Pods {
   /**
