+++ src/core/network_node_services.ts
@@ -3,7 +3,7 @@
  */
 
 import {type NodeAlias} from '../types/aliases.js';
-import {type PodName} from './kube/pod_name.js';
+import {type PodName} from './kube/resources/pod/pod_name.js';
 import {type NamespaceName} from './kube/resources/namespace/namespace_name.js';
 
 export class NetworkNodeServices {
