+++ src/core/platform_installer.ts
@@ -20,7 +20,7 @@ import {sleep} from './helpers.js';
 import {inject, injectable} from 'tsyringe-neo';
 import {patchInject} from './container_helper.js';
 import {type NamespaceName} from './kube/resources/namespace/namespace_name.js';
-import {type PodRef} from './kube/pod_ref.js';
+import {type PodRef} from './kube/resources/pod/pod_ref.js';
 import {ContainerRef} from './kube/container_ref.js';
 import {SecretType} from './kube/secret_type.js';
