+++ src/core/templates.ts
@@ -8,7 +8,7 @@ import {DataValidationError, SoloError, IllegalArgumentError, MissingArgumentErr
 import * as constants from './constants.js';
 import {type AccountId} from '@hashgraph/sdk';
 import {type IP, type NodeAlias, type NodeId} from '../types/aliases.js';
-import {PodName} from './kube/pod_name.js';
+import {PodName} from './kube/resources/pod/pod_name.js';
 import {GrpcProxyTlsEnums} from './enumerations.js';
 import {HEDERA_PLATFORM_VERSION} from '../../version.js';
 import {type NamespaceName} from './kube/resources/namespace/namespace_name.js';
