+++ test/e2e/integration/core/platform_installer_e2e.test.ts
@@ -15,8 +15,8 @@ import {type K8} from '../../../../src/core/kube/k8.js';
 import {type AccountManager} from '../../../../src/core/account_manager.js';
 import {type PlatformInstaller} from '../../../../src/core/platform_installer.js';
 import {NamespaceName} from '../../../../src/core/kube/resources/namespace/namespace_name.js';
-import {PodName} from '../../../../src/core/kube/pod_name.js';
-import {PodRef} from '../../../../src/core/kube/pod_ref.js';
+import {PodName} from '../../../../src/core/kube/resources/pod/pod_name.js';
+import {PodRef} from '../../../../src/core/kube/resources/pod/pod_ref.js';
 import {ContainerRef} from '../../../../src/core/kube/container_ref.js';
 
 const defaultTimeout = Duration.ofSeconds(20).toMillis();
