+++ test/test_util.ts
@@ -46,7 +46,7 @@ import {Duration} from '../src/core/time/duration.js';
 import {container} from 'tsyringe-neo';
 import {resetForTest} from './test_container.js';
 import {NamespaceName} from '../src/core/kube/resources/namespace/namespace_name.js';
-import {PodRef} from '../src/core/kube/pod_ref.js';
+import {PodRef} from '../src/core/kube/resources/pod/pod_ref.js';
 import {ContainerRef} from '../src/core/kube/container_ref.js';
 import {NetworkNodes} from '../src/core/network_nodes.js';
